requirements.txt
flywheel-sdk==8.3.0
google-oauth==1.0.1
google-cloud-automl==0.2.0
google-cloud-storage==1.16.1
pydicom==1.1.0
med2image==1.1.2
pystache==0.5.4
shortuuid==0.5.0
git+https://github.com/flywheel-io/healthcare-api@develop
